convert bin/facet_coverage_check to a unit test
